change ref to user navigation
